Flyttet aktør og bruker til en ny pakke, aktør
