+++ modell/src/main/kotlin/no/nav/tilbakekreving/behov/IverksettelseBehov.kt
@@ -1,8 +1,8 @@
 package no.nav.tilbakekreving.behov
 
+import no.nav.tilbakekreving.aktør.Aktør
 import no.nav.tilbakekreving.beregning.delperiode.Delperiode
 import no.nav.tilbakekreving.fagsystem.Ytelsestype
-import no.nav.tilbakekreving.hendelse.KravgrunnlagHendelse
 import no.nav.tilbakekreving.kontrakter.behandling.Behandlingstype
 import java.util.UUID
 
@@ -12,6 +12,6 @@ class IverksettelseBehov(
     val delperioder: List<Delperiode<out Delperiode.Beløp>>,
     val ansvarligSaksbehandler: String,
     val ytelsestype: Ytelsestype,
-    val aktør: KravgrunnlagHendelse.Aktør,
+    val aktør: Aktør,
     val behandlingstype: Behandlingstype,
 ) : Behov
