+++ modell/src/main/kotlin/no/nav/tilbakekreving/hendelse/FagsysteminfoHendelse.kt
@@ -1,9 +1,10 @@
 package no.nav.tilbakekreving.hendelse
 
+import no.nav.tilbakekreving.aktør.Aktør
 import java.time.LocalDate
 
 data class FagsysteminfoHendelse(
-    val aktør: KravgrunnlagHendelse.Aktør,
+    val aktør: Aktør,
     val behandlingId: String,
     val revurderingsresultat: String,
     val revurderingsårsak: String,
