+++ modell/src/test/kotlin/no/nav/tilbakekreving/hendelse/KravgrunnlagHendelseTest.kt
@@ -3,6 +3,7 @@ package no.nav.tilbakekreving.hendelse
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import no.nav.tilbakekreving.UtenforScope
+import no.nav.tilbakekreving.aktør.Aktør
 import no.nav.tilbakekreving.feil.UtenforScopeException
 import no.nav.tilbakekreving.kravgrunnlag
 import org.junit.jupiter.api.Test
@@ -12,8 +13,8 @@ class KravgrunnlagHendelseTest {
     fun `kravgrunnlag hvor mottaker er ulik bruker er utenfor scope`() {
         val exception = shouldThrow<UtenforScopeException> {
             kravgrunnlag(
-                vedtakGjelder = KravgrunnlagHendelse.Aktør.Person("04056912345"),
-                utbetalesTil = KravgrunnlagHendelse.Aktør.Person("20046912345"),
+                vedtakGjelder = Aktør.Person("04056912345"),
+                utbetalesTil = Aktør.Person("20046912345"),
             )
         }
 
@@ -24,7 +25,7 @@ class KravgrunnlagHendelseTest {
     fun `kan ikke håndtere kravgrunnlag som ikke gjelder person`() {
         val exception = shouldThrow<UtenforScopeException> {
             kravgrunnlag(
-                vedtakGjelder = KravgrunnlagHendelse.Aktør.Organisasjon("889640782"),
+                vedtakGjelder = Aktør.Organisasjon("889640782"),
             )
         }
